Tamanho registro para u32

+++ program/src/state/certificate_info.rs
@@ -4,7 +4,7 @@ use borsh::{ BorshDeserialize, BorshSerialize };
 #[derive(BorshDeserialize, BorshSerialize, Debug)]
 pub struct CertificateInfo {
     pub name: String,
-    pub registration: u8,
+    pub registration: u32,
     pub state: String,
 }
 
@@ -12,7 +12,7 @@ impl CertificateInfo {
 
     pub fn new(
         name: String,
-        registration: u8,
+        registration: u32,
         state: String,
     ) -> Self {
         CertificateInfo {
